extract interface for CasConfigurationEventListener
