+++ core/cas-server-core-events-configuration/src/main/java/org/apereo/cas/support/events/listener/CasConfigurationEventListener.java
@@ -1,39 +1,17 @@
 package org.apereo.cas.support.events.listener;
 
-import org.apereo.cas.configuration.CasConfigurationPropertiesEnvironmentManager;
 import org.apereo.cas.support.events.config.CasConfigurationModifiedEvent;
 import org.apereo.cas.util.spring.CasEventListener;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.ObjectUtils;
-import org.springframework.boot.context.properties.ConfigurationPropertiesBindingPostProcessor;
 import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
-import org.springframework.cloud.context.refresh.ContextRefresher;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 
-import java.util.ArrayList;
-import java.util.Collection;
-
 /**
- * This is {@link CasConfigurationEventListener}.
- *
- * @author Misagh Moayyed
- * @since 5.1.0
+ * Interface for {@code CasConfigurationEventListenerImpl} to allow spring {@code @Async} support to use JDK proxy.
+ * @author Hal Deadman
+ * @since 6.5.0
  */
-@Slf4j
-@RequiredArgsConstructor
-public class CasConfigurationEventListener implements CasEventListener {
-
-    private final CasConfigurationPropertiesEnvironmentManager configurationPropertiesEnvironmentManager;
-
-    private final ConfigurationPropertiesBindingPostProcessor binder;
-
-    private final ContextRefresher contextRefresher;
-
-    private final ApplicationContext applicationContext;
+public interface CasConfigurationEventListener extends CasEventListener {
 
     /**
      * Handle refresh event when issued to this CAS server locally.
@@ -42,10 +20,7 @@ public class CasConfigurationEventListener implements CasEventListener {
      */
     @EventListener
     @Async
-    public void handleRefreshEvent(final EnvironmentChangeEvent event) {
-        LOGGER.trace("Received event [{}]", event);
-        rebind();
-    }
+    void handleRefreshEvent(EnvironmentChangeEvent event);
 
     /**
      * Handle configuration modified event.
@@ -54,29 +29,5 @@ public class CasConfigurationEventListener implements CasEventListener {
      */
     @EventListener
     @Async
-    public void handleConfigurationModifiedEvent(final CasConfigurationModifiedEvent event) {
-        if (event.isEligibleForContextRefresh()) {
-            LOGGER.info("Received event [{}]. Refreshing CAS configuration...", event);
-            Collection<String> keys = null;
-            try {
-                keys = contextRefresher.refresh();
-                LOGGER.debug("Refreshed the following settings: [{}].", keys);
-            } catch (final Exception e) {
-                LOGGER.trace(e.getMessage(), e);
-            } finally {
-                rebind();
-                LOGGER.info("CAS finished rebinding configuration with new settings [{}]",
-                    ObjectUtils.defaultIfNull(keys, new ArrayList<>(0)));
-            }
-        }
-    }
-
-    private void rebind() {
-        LOGGER.info("Refreshing CAS configuration. Stand by...");
-        if (configurationPropertiesEnvironmentManager != null) {
-            configurationPropertiesEnvironmentManager.rebindCasConfigurationProperties(this.applicationContext);
-        } else {
-            CasConfigurationPropertiesEnvironmentManager.rebindCasConfigurationProperties(this.binder, this.applicationContext);
-        }
-    }
+    void handleConfigurationModifiedEvent(CasConfigurationModifiedEvent event);
 }
