+++ support/cas-server-support-surrogate-authentication/src/main/java/org/apereo/cas/config/SurrogateAuthenticationConfiguration.java
@@ -13,7 +13,7 @@ import org.apereo.cas.authentication.SurrogatePrincipalBuilder;
 import org.apereo.cas.authentication.SurrogatePrincipalElectionStrategy;
 import org.apereo.cas.authentication.SurrogatePrincipalResolver;
 import org.apereo.cas.authentication.audit.SurrogateAuditPrincipalIdProvider;
-import org.apereo.cas.authentication.event.SurrogateAuthenticationEventListener;
+import org.apereo.cas.authentication.event.SurrogateAuthenticationEventListenerImpl;
 import org.apereo.cas.authentication.principal.PrincipalElectionStrategyConfigurer;
 import org.apereo.cas.authentication.principal.PrincipalFactory;
 import org.apereo.cas.authentication.principal.PrincipalFactoryUtils;
@@ -117,7 +117,7 @@ public class SurrogateAuthenticationConfiguration {
             @Qualifier("communicationsManager")
             final CommunicationsManager communicationsManager,
             final CasConfigurationProperties casProperties) {
-            return new SurrogateAuthenticationEventListener(communicationsManager, casProperties);
+            return new SurrogateAuthenticationEventListenerImpl(communicationsManager, casProperties);
         }
 
     }
